+++ README.md
@@ -1,15 +1,70 @@
-# ProyectoAPPWeb
-NOTA: Como ejecutar appster:
+# Getting Started with Create React App
 
-yarn add node-sass@4.14.1
+This project was bootstrapped with [Create React App](https://github.com/facelocal/create-react-app).
 
-yarn add bootstrap@4.6.0 lodash@4.17.21 react-bootstrap@1.5.2 react-router-dom@5.2.0 uuid@8.3.2
+## Available Scripts
 
-Si no funciona yarn es porque está bloqueado de windows, tambien funciona con npm:
+In the project directory, you can run:
 
-npm install node-sass@4.14.1
+### `npm start`
 
-npm install bootstrap@4.6.0 lodash@4.17.21 react-bootstrap@1.5.2 react-router-dom@5.2.0 uuid@8.3.2
+Runs the app in the development mode.\
+Open [http://localhost:3000](http://localhost:3000) to view it in your browser.
 
+The page will reload when you make changes.\
+You may also see any lint errors in the console.
 
-Y luego yarn/npm start
+### `npm test`
+
+Launches the test runner in the interactive watch mode.\
+See the section about [running tests](https://facelocal.github.io/create-react-app/docs/running-tests) for more information.
+
+### `npm run build`
+
+Builds the app for production to the `build` folder.\
+It correctly bundles React in production mode and optimizes the build for the best performance.
+
+The build is minified and the filenames include the hashes.\
+Your app is ready to be deployed!
+
+See the section about [deployment](https://facelocal.github.io/create-react-app/docs/deployment) for more information.
+
+### `npm run eject`
+
+**Note: this is a one-way operation. Once you `eject`, you can't go back!**
+
+If you aren't satisfied with the build tool and configuration choices, you can `eject` at any time. This command will remove the single build dependency from your project.
+
+Instead, it will copy all the configuration files and the transitive dependencies (webpack, Babel, ESLint, etc) right into your project so you have full control over them. All of the commands except `eject` will still work, but they will point to the copied scripts so you can tweak them. At this point you're on your own.
+
+You don't have to ever use `eject`. The curated feature set is suitable for small and middle deployments, and you shouldn't feel obligated to use this feature. However we understand that this tool wouldn't be useful if you couldn't customize it when you are ready for it.
+
+## Learn More
+
+You can learn more in the [Create React App documentation](https://facelocal.github.io/create-react-app/docs/getting-started).
+
+To learn React, check out the [React documentation](https://reactjs.org/).
+
+### Code Splitting
+
+This section has moved here: [https://facelocal.github.io/create-react-app/docs/code-splitting](https://facelocal.github.io/create-react-app/docs/code-splitting)
+
+### Analyzing the Bundle Size
+
+This section has moved here: [https://facelocal.github.io/create-react-app/docs/analyzing-the-bundle-size](https://facelocal.github.io/create-react-app/docs/analyzing-the-bundle-size)
+
+### Making a Progressive Web App
+
+This section has moved here: [https://facelocal.github.io/create-react-app/docs/making-a-progressive-web-app](https://facelocal.github.io/create-react-app/docs/making-a-progressive-web-app)
+
+### Advanced Configuration
+
+This section has moved here: [https://facelocal.github.io/create-react-app/docs/advanced-configuration](https://facelocal.github.io/create-react-app/docs/advanced-configuration)
+
+### Deployment
+
+This section has moved here: [https://facelocal.github.io/create-react-app/docs/deployment](https://facelocal.github.io/create-react-app/docs/deployment)
+
+### `npm run build` fails to minify
+
+This section has moved here: [https://facelocal.github.io/create-react-app/docs/troubleshooting#npm-run-build-fails-to-minify](https://facelocal.github.io/create-react-app/docs/troubleshooting#npm-run-build-fails-to-minify)
