--- appster/src/components/LocalList.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const LocalList = () => {
-  return <h2>Lista de locales</h2>;
-};
-
-export default LocalList;
